#13049 Refactor RecordTransformer to include RecordEnricher. Removed RecordEnricher as it serves similar purpose as RecordTransformer.
