+++ pinot-segment-local/src/main/java/org/apache/pinot/plugin/record/enricher/RecordEnricherPipeline.java
@@ -16,13 +16,14 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.spi.recordenricher;
+package org.apache.pinot.plugin.record.enricher;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import org.apache.pinot.segment.local.recordtransformer.RecordTransformer;
 import org.apache.pinot.spi.config.table.TableConfig;
 import org.apache.pinot.spi.config.table.ingestion.EnrichmentConfig;
 import org.apache.pinot.spi.config.table.ingestion.IngestionConfig;
@@ -30,7 +31,7 @@ import org.apache.pinot.spi.data.readers.GenericRow;
 
 
 public class RecordEnricherPipeline {
-  private final List<RecordEnricher> _enrichers = new ArrayList<>();
+  private final List<RecordTransformer> _enrichers = new ArrayList<>();
   private final Set<String> _columnsToExtract = new HashSet<>();
 
   public static RecordEnricherPipeline getPassThroughPipeline() {
@@ -45,7 +46,7 @@ public class RecordEnricherPipeline {
     List<EnrichmentConfig> enrichmentConfigs = ingestionConfig.getEnrichmentConfigs();
     for (EnrichmentConfig enrichmentConfig : enrichmentConfigs) {
       try {
-        RecordEnricher enricher = RecordEnricherRegistry.createRecordEnricher(enrichmentConfig);
+        RecordTransformer enricher = RecordEnricherRegistry.createRecordEnricher(enrichmentConfig);
         pipeline.add(enricher);
       } catch (IOException e) {
         throw new RuntimeException("Failed to instantiate record enricher " + enrichmentConfig.getEnricherType(), e);
@@ -62,14 +63,14 @@ public class RecordEnricherPipeline {
     return _columnsToExtract;
   }
 
-  public void add(RecordEnricher enricher) {
+  public void add(RecordTransformer enricher) {
     _enrichers.add(enricher);
     _columnsToExtract.addAll(enricher.getInputColumns());
   }
 
   public void run(GenericRow record) {
-    for (RecordEnricher enricher : _enrichers) {
-      enricher.enrich(record);
+    for (RecordTransformer enricher : _enrichers) {
+      enricher.transform(record);
     }
   }
 }
