+++ pinot-segment-local/src/main/java/org/apache/pinot/plugin/record/enricher/function/CustomFunctionEnricher.java
@@ -27,15 +27,15 @@ import java.util.List;
 import java.util.Map;
 import org.apache.pinot.segment.local.function.FunctionEvaluator;
 import org.apache.pinot.segment.local.function.FunctionEvaluatorFactory;
+import org.apache.pinot.segment.local.recordtransformer.RecordTransformer;
 import org.apache.pinot.spi.data.readers.GenericRow;
-import org.apache.pinot.spi.recordenricher.RecordEnricher;
 import org.apache.pinot.spi.utils.JsonUtils;
 
 
 /**
  * Enriches the record with custom functions.
  */
-public class CustomFunctionEnricher implements RecordEnricher {
+public class CustomFunctionEnricher implements RecordTransformer {
   private final Map<String, FunctionEvaluator> _fieldToFunctionEvaluator;
   private final List<String> _fieldsToExtract;
 
@@ -58,9 +58,10 @@ public class CustomFunctionEnricher implements RecordEnricher {
   }
 
   @Override
-  public void enrich(GenericRow record) {
+  public GenericRow transform(GenericRow record) {
     _fieldToFunctionEvaluator.forEach((field, evaluator) -> {
       record.putValue(field, evaluator.evaluate(record));
     });
+    return record;
   }
 }
