+++ pinot-segment-local/src/main/java/org/apache/pinot/plugin/record/enricher/function/CustomFunctionEnricherFactory.java
@@ -22,9 +22,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.google.auto.service.AutoService;
 import java.io.IOException;
 import org.apache.pinot.segment.local.function.FunctionEvaluatorFactory;
-import org.apache.pinot.spi.recordenricher.RecordEnricher;
-import org.apache.pinot.spi.recordenricher.RecordEnricherFactory;
-import org.apache.pinot.spi.recordenricher.RecordEnricherValidationConfig;
+import org.apache.pinot.plugin.record.enricher.RecordEnricherFactory;
+import org.apache.pinot.plugin.record.enricher.RecordEnricherValidationConfig;
+import org.apache.pinot.segment.local.recordtransformer.RecordTransformer;
 import org.apache.pinot.spi.utils.JsonUtils;
 
 @AutoService(RecordEnricherFactory.class)
@@ -36,7 +36,7 @@ public class CustomFunctionEnricherFactory implements RecordEnricherFactory {
   }
 
   @Override
-  public RecordEnricher createEnricher(JsonNode enricherProps)
+  public RecordTransformer createEnricher(JsonNode enricherProps)
       throws IOException {
     return new CustomFunctionEnricher(enricherProps);
   }
