+++ pinot-segment-local/src/main/java/org/apache/pinot/segment/local/recordtransformer/RecordTransformer.java
@@ -19,6 +19,7 @@
 package org.apache.pinot.segment.local.recordtransformer;
 
 import java.io.Serializable;
+import java.util.List;
 import javax.annotation.Nullable;
 import org.apache.pinot.spi.data.readers.GenericRow;
 
@@ -43,4 +44,10 @@ public interface RecordTransformer extends Serializable {
    */
   @Nullable
   GenericRow transform(GenericRow record);
+
+  /**
+   * Returns the list of input columns required for enriching the record.
+   * This is used to make sure the required input fields are extracted.
+   */
+  List<String> getInputColumns();
 }
